Added create/update/delete for all entities.

+++ src/main/java/ord/example/LegalManagementSystem/controller/ClientController.java
@@ -1,6 +1,7 @@
 package ord.example.LegalManagementSystem.controller;
 
 import jakarta.validation.Valid;
+import lombok.extern.slf4j.Slf4j;
 import ord.example.LegalManagementSystem.dtos.Client.ClientCUDTO;
 import ord.example.LegalManagementSystem.dtos.Client.ClientReadDTO;
 import ord.example.LegalManagementSystem.exceptions.ClientNotFoundException;
@@ -15,6 +16,7 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 import java.util.Optional;
 
+@Slf4j
 @Controller
 @RequestMapping("/clients")
 public class ClientController {
@@ -25,20 +27,21 @@ public class ClientController {
     }
 
     @GetMapping("/create")
-    public String getClientForm(Model model) {
+    public String getCreteClientForm(Model model) {
         model.addAttribute("client", new ClientCUDTO());
+        model.addAttribute("formAction", "/clients");
         return "client/form";
     }
 
     @PostMapping
     public String createClient(@Valid @ModelAttribute("client") ClientCUDTO clientCUDTO, BindingResult bindingResult) {
-        ClientReadDTO savedClient = clientService.saveClient(clientCUDTO);
-
         if (bindingResult.hasErrors()) {
             return "client/form";
-        } else {
-            return "redirect:/clients/" + savedClient.getClientId();
         }
+
+        ClientReadDTO savedClient = clientService.saveClient(clientCUDTO);
+
+        return "redirect:/clients/" + savedClient.getClientId();
     }
 
     @GetMapping("/{clientId}")
@@ -61,20 +64,45 @@ public class ClientController {
         return "client/clients";
     }
 
-    @PutMapping("/{clientId}")
-    public ResponseEntity<ClientReadDTO> updateClient(@PathVariable Integer clientId, @Valid @RequestBody ClientCUDTO clientCUDTO) {
-        Optional<ClientReadDTO> updatedClient = clientService.updateClientById(clientId, clientCUDTO);
-        return updatedClient.map(c -> new ResponseEntity<>(c, HttpStatus.OK))
-                .orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
+    @GetMapping("/edit/{clientId}")
+    public String getEditClientForm(@PathVariable Integer clientId, Model model) {
+        Optional<ClientReadDTO> client = clientService.getClientById(clientId);
+
+        if (client.isEmpty()) {
+            throw new ClientNotFoundException("Client with ID " + clientId + " not found");
+        } else {
+            model.addAttribute("client", client.get());
+            model.addAttribute("formAction", "/clients/" + clientId);
+            return "client/form";
+        }
     }
 
-    @DeleteMapping("/{clientId}")
-    public ResponseEntity<Void> deleteClient(@PathVariable Integer clientId) {
-        try {
+    @PostMapping("/{clientId}")
+    public String updateClient(@PathVariable Integer clientId, @Valid @ModelAttribute("client") ClientCUDTO clientCUDTO,
+                               BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return "client/form";
+        }
+
+        Optional<ClientReadDTO> updatedClient = clientService.updateClientById(clientId, clientCUDTO);
+
+        if (updatedClient.isEmpty()) {
+            throw new ClientNotFoundException("Client with ID " + clientId + " not found");
+        }
+
+        return "redirect:/clients/" + updatedClient.get().getClientId();
+    }
+
+
+    @PostMapping("/delete/{clientId}")
+    public String deleteClient(@PathVariable Integer clientId) {
+        Optional<ClientReadDTO> client = clientService.getClientById(clientId);
+
+        if (client.isPresent()) {
             clientService.deleteClientById(clientId);
-            return new ResponseEntity<>(HttpStatus.OK);
-        } catch (Exception ex) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            return "redirect:/clients";
+        } else {
+            throw new ClientNotFoundException("Client with ID " + clientId + " not found");
         }
     }
 }
